Extract panel visibility & resize hooks
